Moved yield and exit to WorkflowThread
